+++ Asset.ViewModels/HospitalVM/DetailHospitalVM.cs
@@ -30,9 +30,9 @@ namespace Asset.ViewModels.HospitalVM
         public string ManagerNameAr { get; set; }
 
 
-        public double? Latitude { get; set; }
+        public decimal? Latitude { get; set; }
 
-        public double? Longtitude { get; set; }
+        public decimal? Longtitude { get; set; }
 
         public string Address { get; set; }
